Pull Table of OUs/Countries, UIDs, ISO codes and levels

`get_outable` pulls from DATIM to return a dataframe with all PEPFAR Operating Units and countries along with useful information for merging, eg ISO codes, and use in DATIM APIs, eg UIDs and hierarchy levels.

get_outable(username, password, baseurl = "https://final.datim.org/")

Arguments

username
DATIM Username, defaults to using `datim_user()` if blank

password
DATIM password, defaults to using `datim_pwd()` if blank

baseurl
base url for the API, default = https://final.datim.org/

Value

data frame with all PEPFAR OUs, countries, their UIDs, ISO codes and different levels in the DATIM hierarchy

Details

`get_outtable` is a wrapper around `identify_ouuids` and `identify_levels` that pulls this information directly from DATIM. The user will need to have a DATIM account to access this data. You can take advantage of storing you credentials locally in a secure way using `set_datim`.

See also

Examples

if (FALSE) {
 load_secrets()
 ou_table <- datim_outable() }