Cross-walk of Country names

A dataset PEPFAR Operating Units and Countries along with their ISO codes, alternative names from other sources. This is a useful dataset designed to help with data cleaning / matching from different sources.

pepfar_country_xwalk

Format

A data frame with 55 rows and 29 variables:

iso3c

ISO-3 Code

continent

Continent name

region

World Region Name

region23

Alternative name for World Region

un_region_name

US Region Name

iso_name_en

ISO Country name in English

iso_name_fr

SO Country name in French

country_name_de

Regular Expression of Country name in German

country_name_de_regex

Country name in German

country_name_en

Country name in English

country_name_en_regex

Regular Expression of Country name in English

country_name_fr

Regular Expression of Country name in French

country_name_fr_regex

Country name in French

country_name_it

Regular Expression of Country name in Italian

country_name_it_regex

Country name in Italian

un_name_ar

UN Name in Arabic

un_name_en

UN Name in English

un_name_es

UN Name in Spanish

un_name_fr

UN Name in French

un_name_ru

UN Name in ru

un_name_zh

UN Name in zh

sovereignt

rnaturalearth sovereign territory name

admin

rnaturalearth administrative unit name

name

rnaturalearth country name

operatingunit

PEPFAR Operating Unit (countries + 3 regional programs)

operatingunit_iso

PEPFAR Operating Unit ISO-3

operatingunit_uid

PEPFAR Operating Unit unique id from DATIM

country

PEPFAR Country Name

country_uid

PEPFAR Country unique id from DATIM

Source

https://final.datim.org/ https://www.naturalearthdata.com/ https://vincentarelbundock.github.io/countrycode/