Resolve Known Issues

This function handles known issues in the MSD. For example, a mechanism starting mid-year and reporting on TX_CURR will duplicate TX_CURR targets. This function resolves that by removing those known cases and stores cases relevant to your data frame.

The best workflow is to filter your dataset down to a country and/or technical area of interest before running `resolve_knownissues()`. When you run the function, it will print out any known issues to the console (and can) even store them to your Global Environment, so it makes sense to limit the data first to what you are using/care about.

The list of known issues is maintained by USAID/SIEI division. To access the table requires having a USAID email account and can be accessed via `browse_knownissues()`.

resolve_knownissues(df, remove_cs = TRUE, store_excl = FALSE)

Arguments

df

standard MSD data frame, typically after its been filtered

remove_cs

remove data flagged as central support (CS), default = TRUE

store_excl

should the known exclusions be store in the Global Envir?

Value

df excluding known targets/results issues

See also

Examples

if (FALSE) {

library(tidyverse)
library(glamr)

load_secrets() # or googlesheets4::gs4_auth()

df_msd <- si_path() %>%
  return_latest("OU_IM") %>%
  read_rds()

df_mwi <- df_msd %>%
  filter(operatingunit == "Malawi",
         indicator == "TX_CURR")

df_mwi_resolved <- df_mwi %>%
  resolve_knownissues() }